Add back CMake build config for macOS (#2729)

+++ platform/glfw/glfw_view.cpp
@@ -218,7 +218,7 @@ GLFWView::GLFWView(bool fullscreen_,
 #ifdef __APPLE__
     int fbW, fbH;
     glfwGetFramebufferSize(window, &fbW, &fbH);
-    backend->setSize({fbW, fbH});
+    backend->setSize({static_cast<uint32_t>(fbW), static_cast<uint32_t>(fbH)});
 #endif
 
     pixelRatio = static_cast<float>(backend->getSize().width) / width;
@@ -900,7 +900,7 @@ void GLFWView::onWindowResize(GLFWwindow *window, int width, int height) {
 #ifdef __APPLE__
     int fbW, fbH;
     glfwGetFramebufferSize(window, &fbW, &fbH);
-    view->backend->setSize({fbW, fbH});
+    view->backend->setSize({static_cast<uint32_t>(fbW), static_cast<uint32_t>(fbH)});
 #endif
 }
